Fix STORE_AUDIT.md T10: widen UndoId from u16 to u64
UndoId was minted as id.len() as u16, where id.len() is a transaction's
total undo-op count across every row it touches — wrapping at 65,536
ops. A wrapped id then aliases find_undo_tuple's per-txn Vec index,
handing a concurrent MVCC read a different row's pre-image entirely.

Widened to u64 (not the audit's stated minimum of u32) since the
planned T4+S2 WAL redesign will likely make the undo pointer be the
record's own LsnId (already u64), avoiding a second width change later.

The audit's own suggested repro (one row, 70k updates, then rollback)
doesn't actually exercise the bug — rollback replays the raw undo Vec
directly, never through UndoId, and repeated updates to one row always
resolve to an identical pre-image regardless of slot. Test instead
drives the bug's other named consequence: a wrapped undo_id aliasing a
different row's undo entry during a concurrent MVCC read.

T4+S2 (single framed/checksummed WAL) remain staged as a separate
follow-up per audit-progress.md.

+++ store/src/logger.rs
@@ -83,7 +83,7 @@ impl LsnClock {
 pub struct LsnId(pub(crate) u64);
 
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Hash, Serialize, Deserialize)]
-pub struct UndoId(pub(crate) u16);
+pub struct UndoId(pub(crate) u64);
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) enum MsgType {
@@ -207,7 +207,7 @@ impl Logger {
                     .and_modify(|v| v.push(op.clone()))
                     .or_insert(vec![op.clone()]);
                 MsgType::Undo(UndoOperation {
-                    undo_id: Some(UndoId(id.len() as u16)),
+                    undo_id: Some(UndoId(id.len() as u64)),
                     operation: op.clone(),
                 })
             }
@@ -597,7 +597,7 @@ fn redo_log_runner(
 
 impl From<usize> for UndoId {
     fn from(value: usize) -> Self {
-        Self(value as u16)
+        Self(value as u64)
     }
 }
